PR fixes to use utils
