+++ my-app/src/app/square.js
